python-agents/pyproject.toml
[project]
name = "absher-visa-agents"
version = "1.0.0"
description = "AI-powered visa application assistant agents for Absher"
requires-python = ">=3.11"
authors = [
    { name = "Absher Team" }
]
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # AI/LLM - LangChain ecosystem
    "langchain>=0.1.0",
    "langgraph>=0.0.26",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "ollama>=0.1.6",

    # Embeddings & Vector DB
    "sentence-transformers>=2.3.0",
    "pgvector>=0.2.4",

    # Web Scraping
    "crawl4ai>=0.2.0",
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.26.0",

    # OCR for Arabic document processing
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",

    # PDF Generation with Arabic RTL support
    "weasyprint>=60.2",
    "jinja2>=3.1.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Caching
    "redis>=5.0.0",
    "hiredis>=2.3.0",

    # Storage
    "boto3>=1.34.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
